feat(member): add CommunityPostsSection Component (#119)
